requirements.txt
apache-airflow-providers-google==10.16.0
dbt-core==1.7.10
dbt-bigquery==1.7.6